* Various optimizations to the XDAG pool startup time
**
Removed much of the logging when reloading the old blocks from storage.
As the logging also did a disk flush it was intrusive for disk activiy on my system (other Pools milage might differ).

* Various tweaks that can be used to imrpove XDAG pool startup time and running performance

**
-z <path>
Command line argument added, to specify a locaiton for the temporary file contating "pool memory swap".

**
-z malloc
Command line argument added, this removes the "pool memory swap" completly, and the pool relies on the OS capabilities
for this if required.

* Changed whitelist POOL IP from 52.191.198.174 to 40.71.5.58 , this is the machine where all this development
and testing has happened.

Future things...

Better ways for error logging is probobly a good thing.
Replace command line argument with a config file design is probobly a good thing.

I would appritiate to see various profilings of Pool starups to assist in where the bottle necks are.